+++ server/data/repository.go
@@ -7,7 +7,7 @@ type Repository interface {
 	InsertBuyOrder(order BuyOrder, breads []Bread) (int, error)
 	InsertMakeOrder(order MakeOrder, breads []Bread) (int, error)
 	AdjustBreadQuantity(breadID int, quantityChange int) error
-	AdjustBreadPrice(breadID int, newPrice float64) error
+	AdjustBreadPrice(breadID int, newPrice float32) error
 	PasswordMatches(plainText string, customer Customer) (bool, error)
 	GetAvailableBread() ([]Bread, error)
 	GetBreadByID(breadID int) (bread Bread, err error)
